split reshape transforms into separate files
